remove AbstractResultService
