+++ models/classes/AbstractResultService.php
@@ -29,57 +29,4 @@ use \taoResultServer_models_classes_WritableResultStorage as WritableResultStora
 abstract class AbstractResultService extends ConfigurableService implements ResultServerService
 {
 
-    /**
-     * Starts or resume a taoResultServerStateFull session for results submission
-     *
-     * @param \core_kernel_classes_Resource $compiledDelivery
-     * @param string $executionIdentifier
-     * @param string $userUri
-     * @throws
-     */
-    public function initResultServer($compiledDelivery, $executionIdentifier, $userUri)
-    {
-        $storage = $this->getResultStorage();
-        //$storage->spawnResult($executionIdentifier);
-
-        //link test taker identifier with results
-        $storage->storeRelatedTestTaker($executionIdentifier, $userUri);
-
-        //link delivery identifier with results
-        $storage->storeRelatedDelivery($executionIdentifier, $compiledDelivery->getUri());
-    }
-
-    /**
-     * @param string $serviceId
-     * @return WritableResultStorage
-     * @throws \common_exception_Error
-     * @throws \oat\oatbox\service\exception\InvalidServiceManagerException
-     */
-    public function instantiateResultStorage($serviceId)
-    {
-        $storage = null;
-        if (class_exists($serviceId)) { //some old serialized session can has class name instead of service id
-            $storage = new $serviceId();
-        } elseif ($this->getServiceManager()->has($serviceId)) {
-            $storage = $this->getServiceManager()->get($serviceId);
-        }
-
-        if ($storage instanceof ServiceLocatorAwareInterface) {
-            $storage->setServiceLocator($this->getServiceLocator());
-        }
-
-        if ($storage === null || !$storage instanceof WritableResultStorage) {
-            throw new \common_exception_Error('Configured result storage is not writable.');
-        }
-
-        return $storage;
-    }
-
-    abstract public function getResultStorage();
-
-    public function deleteDeliveryExecutionData(DeliveryExecutionDeleteRequest $request)
-    {
-        $storage = $this->getResultStorage();
-        return $storage->deleteDeliveryExecutionData($request);
-    }
 }
